finish lab1

+++ src/include/buffer/lru_replacer.h
@@ -49,7 +49,7 @@ class LRUReplacer : public Replacer {
  private:
   // TODO(student): implement me!
   std::mutex latch_;
-  int capacity_;
+  size_t capacity_;
   std::list<frame_id_t> lru_list_;
   std::unordered_map<frame_id_t, std::list<frame_id_t>::iterator> lru_map_;
 };
